EPMRPP-104683: moved ExecutionStatusPopover to shared manualLaunches components
