+++ app/src/pages/inside/manualLaunchesPage/executionStatusPopover/executionStatusPopover.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@use '../../../../../common/css/mixins' as mixins;
+@use 'common/css/mixins' as mixins;
 
 .popover {
   padding: 8px 0;
