+++ app/src/pages/inside/manualLaunchesPage/index.js
@@ -1,1 +1,3 @@
 export { ManualLaunchesPage } from './manualLaunchesPage';
+export { ExecutionStatusPopover } from './executionStatusPopover';
+export * from './constants';
