+++ app/src/pages/inside/manualLaunchesPage/manualLaunchExecutionPage/executionStatusButtons/executionStatusButtons.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@use '../../../../../common/css/mixins' as mixins;
+@use 'common/css/mixins.scss' as mixins;
 
 .execution-status-buttons {
   display: flex;
